pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Website analysis application"
autors = [{name = 'Habited', email = 'pdskate88@gmail.com'}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.11",
]

[project.scripts]
page_analyzer = "page_analyzer.app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]
